fix: LendListRepository의 JpaRepository ID 타입 변경
- rent 엔티티가 BaseEntity를 상속받도록 구현된 부분이 적용되지 않았어서 이를 수정

+++ backend/bookbook/src/main/java/com/bookbook/domain/lendList/repository/LendListRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LendListRepository extends JpaRepository<Rent, Long> {
+public interface LendListRepository extends JpaRepository<Rent, Integer> {
     
     @Query("SELECT r FROM Rent r WHERE r.lenderUserId = :lenderUserId AND r.rentStatus != 'DELETED'")
     Page<Rent> findByLenderUserId(@Param("lenderUserId") Long lenderUserId, Pageable pageable);
